github: remove constraint for automated build at forks
With the nightly and on-demand test target, a repository constraint
restricts the github action execution from the forks.

However, by default, workflows are disabled in the forked repository.
The action will be performed when the owner of the fork manually
approves it to execute. (And the action will be executed within that
user's own worker.)

Also, scheduled workflows are disabled by default in forks and to be
enabled, user approval will be required. (And the action will be
executed within that user's own worker.)

This commit removes the constraint for the automated build at forks.

Signed-off-by: Daniel T. Lee <danieltimlee@gmail.com>